Introduce Queued Data
With the current implementation, there was no way to reference
queued data. In this commit, the `queued-data` module is introduced
to do just that. When queuing data to be stored in a block, the
`blockchain` module will assign a temporarily ID for the data in
question making it possible to identify and reference a specific
data entry.

In addition to this, this commit also modifies the following
events:
  * `enqueue-block` is changed to `enqueue-data` and instead of
    emitting the data to be stored, it now emits the whole
    `queue-data` object (i.e. include the `queued-data` id).
  * `constructing-block` is changed so that instead of emitting the
    `block` object which the data will be stored in, it now emits
    the `queue-data` object which the block being constructed will
    store the data of.
  * `new-block` is updated to also emit the `queue-data` id which
    the `block` was created for.