Add functionality to include Blocks
With the current implementation, Blocks could only be added through
Queued Data, which restricted the Blockchain to include Blocks that
where generated by other Blockchains. In this commit a new function
addBlock() was added to the public api to allow the inclusion of
Blocks created by other Blockchain objects. This function will
accept a JSON representation of the Block to be added & the Queued
Data ID which the Block was created for.

When adding a new Block, the Blockchain object will do the
following:

  1. Validate the new Block
     * Validation of Block's hash.
     * Validation of Block's previous hash.
     * Validation of Block's index.
  2. Stop the Block generation.
  3. Include the Block in the Blockchain.
  4. Remove the Queued Data which the newly added Block was created
     for from the Blockchain queue.
  5. Resume Block generation.