Add functionality to include externally generated Queued Data
With the current implementation of the blockchain module, it is not
possible to add a Queued Data object that is initialized from
another Blockchain object. Although it is possible to extract the
Queued Data data and include it in both Blockchain objects, each
Queued Data will be having its own ID which may cause some issues
when implementing blockchain-node since when a Block is created the
Queued Data object which the new Block was created for should be
removed from the queue.

In this commit, logic was implemented to include Queued Data with
their IDs.