Store Queue Data ID in Block
With the current implementation, when creating a Block for a Queue
Data, the Block only stores the Queue Data data. With this
approach there is the chance that multiple Blocks will be created
for the same Queue Data.

In this commit the Block module was enhanced so that apart from
the data it also stores the Queue Data ID.